feat(liburing) update to latest version

+++ modules/lwjgl/core/src/main/c/linux/liburing/arch/syscall-defs.h
@@ -11,9 +11,9 @@ static inline int __sys_open(const char *pathname, int flags, mode_t mode)
 	 * Some architectures don't have __NR_open, but __NR_openat.
 	 */
 #ifdef __NR_open
-	return __do_syscall3(__NR_open, pathname, flags, mode);
+	return (int) __do_syscall3(__NR_open, pathname, flags, mode);
 #else
-	return __do_syscall4(__NR_openat, AT_FDCWD, pathname, flags, mode);
+	return (int) __do_syscall4(__NR_openat, AT_FDCWD, pathname, flags, mode);
 #endif
 }
